+++ package.json
@@ -1,15 +1,15 @@
 {
   "dependencies": {
+    "body-parser": "^1.17.2",
     "cheerio": "^1.0.0-rc.2",
     "collect.js": "^3.0.49",
-    "commander": "^2.11.0",
     "express": "^4.15.4",
     "nightmare": "^2.10.0",
     "node-horseman": "^3.3.0",
     "tabletojson": "^0.6.0"
   },
   "bin": {
-    "booking-kereta" : "index.js"
+    "booking-kereta": "index.js"
   },
   "name": "cek-kode-booking-kereta",
   "version": "1.0.0",
